tools: Move memoizers to memoization.py
