+++ devito/tools/utils.py
@@ -2,8 +2,8 @@ from abc import ABC
 import os
 import inspect
 import ctypes
-from collections import Callable, Iterable, OrderedDict, Hashable, Mapping
-from functools import partial, reduce
+from collections import Callable, Iterable, OrderedDict, Mapping
+from functools import reduce
 from itertools import chain, combinations, product, zip_longest
 from operator import attrgetter, mul
 
@@ -18,8 +18,7 @@ __all__ = ['prod', 'as_tuple', 'is_integer', 'generator', 'grouper', 'split',
            'filter_sorted', 'build_dependence_lists', 'toposort', 'numpy_to_ctypes',
            'ctypes_to_C', 'ctypes_pointer', 'pprint', 'DefaultOrderedDict',
            'change_directory', 'GenericVisitor', 'Bunch',
-           'EnrichedTuple', 'ReducerMap', 'Tag',
-           'memoized_func', 'memoized_meth', 'sweep']
+           'EnrichedTuple', 'ReducerMap', 'Tag', 'sweep']
 
 
 def prod(iterable):
@@ -298,93 +297,6 @@ class change_directory(object):
 
     def __exit__(self, etype, value, traceback):
         os.chdir(self.savedPath)
-
-
-class memoized_func(object):
-    """
-    Decorator. Caches a function's return value each time it is called.
-    If called later with the same arguments, the cached value is returned
-    (not reevaluated). This decorator may also be used on class methods,
-    but it will cache at the class level; to cache at the instance level,
-    use ``memoized_meth``.
-
-    Adapted from: ::
-
-        https://wiki.python.org/moin/PythonDecoratorLibrary#Memoize
-    """
-
-    def __init__(self, func):
-        self.func = func
-        self.cache = {}
-
-    def __call__(self, *args):
-        if not isinstance(args, Hashable):
-            # Uncacheable, a list, for instance.
-            # Better to not cache than blow up.
-            return self.func(*args)
-        if args in self.cache:
-            return self.cache[args]
-        else:
-            value = self.func(*args)
-            self.cache[args] = value
-            return value
-
-    def __repr__(self):
-        """Return the function's docstring."""
-        return self.func.__doc__
-
-    def __get__(self, obj, objtype):
-        """Support instance methods."""
-        return partial(self.__call__, obj)
-
-
-class memoized_meth(object):
-    """
-    Decorator. Cache the return value of a class method.
-
-    Unlike ``memoized_func``, the return value of a given method invocation
-    will be cached on the instance whose method was invoked. All arguments
-    passed to a method decorated with memoize must be hashable.
-
-    If a memoized method is invoked directly on its class the result will not
-    be cached. Instead the method will be invoked like a static method: ::
-
-        class Obj(object):
-            @memoize
-            def add_to(self, arg):
-                return self + arg
-        Obj.add_to(1) # not enough arguments
-        Obj.add_to(1, 2) # returns 3, result is not cached
-
-    Adapted from: ::
-
-        code.activestate.com/recipes/577452-a-memoize-decorator-for-instance-methods/
-    """
-
-    def __init__(self, func):
-        self.func = func
-
-    def __get__(self, obj, objtype=None):
-        if obj is None:
-            return self.func
-        return partial(self, obj)
-
-    def __call__(self, *args, **kw):
-        if not isinstance(args, Hashable):
-            # Uncacheable, a list, for instance.
-            # Better to not cache than blow up.
-            return self.func(*args)
-        obj = args[0]
-        try:
-            cache = obj.__cache
-        except AttributeError:
-            cache = obj.__cache = {}
-        key = (self.func, args[1:], frozenset(kw.items()))
-        try:
-            res = cache[key]
-        except KeyError:
-            res = cache[key] = self.func(*args, **kw)
-        return res
 
 
 def sweep(parameters, keys=None):
